+++ spec/dummy/app/models/dummy_model.rb
@@ -4,4 +4,8 @@ class DummyModel < ApplicationRecord
   belongs_to :asso, polymorphic: true, required: false
 
   polymorphic_id_for :asso
+
+  def state
+    "draft"
+  end
 end